fixed the naming 2
