+++ app/src/main/java/com/movies/movies_and_shows_api/data/remote/MoviesRemoteDataSource.kt
@@ -1,0 +1,10 @@
+package com.movies.movies_and_shows_api.data.remote
+
+
+import com.movies.movies_and_shows_api.domain.Movie
+
+
+interface MoviesRemoteDataSource {
+    fun getMovies(): List<Movie>
+    fun getMovie(movieId: String): Movie?
+}
